pipeline/requirements.txt
numpy>=1.20
pandas>=1.3
PyYAML>=5.4
